requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
idna==3.4
lxml==4.9.2
pdfminer.six==20220524
pycparser==2.21
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13
